Factor JMeterGuiLauncher out of JMeter
